sh: Promote PID storage to int to avoid 16-bit truncation.
PIDs exceeding 32767 were truncated when stored in 'short', causing
mismatches in the job table and infinite busy-waits in waitforjob().
Promoted pid, pgrp, and job counters to int.

+++ jobs.c
@@ -70,10 +70,10 @@ static char sccsid[] = "@(#)jobs.c	8.5 (Berkeley) 5/4/95";
 
 struct job *jobtab;		/* array of jobs */
 int njobs;			/* size of array */
-MKINIT short backgndpid = -1;	/* pid of last background process */
+MKINIT int backgndpid = -1;	/* pid of last background process */
 #if JOBS
 int initialpgrp;		/* pgrp of shell on invocation */
-short curjob;			/* current job */
+int curjob;			/* current job */
 #endif
 
 STATIC void restartjob __P((struct job *));
